fix: details
